backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "authlib>=1.5.2",
    "fastapi>=0.115.11",
    "google-auth>=2.40.1",
    "itsdangerous>=2.2.0",
    "mysql-connector-python>=9.3.0",
    "passlib>=1.7.4",
    "psycopg>=3.2.6",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
    "websockets>=15.0.1",
]
